separando widgets en archivos diferentes
