+++ lib/modules/autenticacion_usuario/screens/register_screen.dart
@@ -1,6 +1,8 @@
 // ignore_for_file: use_build_context_synchronously
 import 'package:flutter/material.dart';
 import 'package:tidal_wave/modules/autenticacion_usuario/classes/tw_user.dart';
+import 'package:tidal_wave/modules/autenticacion_usuario/widgets/auth_text_field.dart';
+import 'package:tidal_wave/modules/autenticacion_usuario/widgets/popup_message.dart';
 import 'package:tidal_wave/services/firebase_auth_service.dart';
 
 class RegisterScreen extends StatefulWidget {
@@ -73,7 +75,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
               //* Username
               Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-                child: RegisterTextField(
+                child: AuthTextField(
                   controller: _usernameController,
                   hintText: 'Nombre de usuario',
                   icon: const Icon(Icons.person_2_rounded),
@@ -96,7 +98,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
               //* Email
               Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-                child: RegisterTextField(
+                child: AuthTextField(
                   controller: _emailController,
                   hintText: 'Email',
                   textInputType: TextInputType.emailAddress,
@@ -113,7 +115,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
               //* Password
               Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-                child: RegisterTextField(
+                child: AuthTextField(
                   controller: _passwordController,
                   hintText: 'Contraseña',
                   textInputType: TextInputType.visiblePassword,
@@ -145,7 +147,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
               //* Verify password
               Padding(
                 padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-                child: RegisterTextField(
+                child: AuthTextField(
                   controller: _verifyPasswordController,
                   hintText: 'Verificar contraseña',
                   textInputType: TextInputType.visiblePassword,
@@ -207,76 +209,5 @@ class _RegisterScreenState extends State<RegisterScreen> {
     _passwordController.dispose();
     _verifyPasswordController.dispose();
     super.dispose();
-  }
-}
-
-class RegisterTextField extends StatelessWidget {
-
-  final String hintText;
-  final TextInputType? textInputType;
-  final Icon? icon;
-  final String? Function(String? value)? validator;
-  final TextEditingController? controller;
-
-  const RegisterTextField({
-    super.key, required this.hintText, this.textInputType, this.icon, this.validator, this.controller
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      validator: validator,
-      controller: controller,
-      keyboardType: textInputType,
-      obscureText: textInputType==TextInputType.visiblePassword,
-      cursorColor: Colors.grey.shade500,
-      autocorrect: false,
-      enableSuggestions: false,
-      style: TextStyle(color: Colors.grey.shade500),
-      decoration: InputDecoration(
-        suffixIcon: icon,
-        suffixIconColor: Colors.grey,
-        filled: true,
-        fillColor: Colors.grey.shade900,
-        hintText: hintText,
-        hintStyle: TextStyle(color: Colors.grey.shade600),
-        enabledBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8.0),
-          borderSide: const BorderSide(color: Colors.transparent)
-        ),
-        focusedBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8.0),
-          borderSide: BorderSide(color: Colors.grey.shade600)
-        ),
-        errorBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8.0),
-          borderSide: BorderSide(color: Colors.red.shade800)
-        ),
-        focusedErrorBorder: OutlineInputBorder(
-          borderRadius: BorderRadius.circular(8.0),
-          borderSide: BorderSide(color: Colors.red.shade400)
-        ),
-      ),
-    );
-  }
-}
-
-class PopupMessage extends StatelessWidget {
-
-  final String title;
-  final String description;
-
-  const PopupMessage({super.key, required this.title, required this.description});
-
-  @override
-  Widget build(BuildContext context) {
-    return AlertDialog(
-      title: Text(title, style: const TextStyle(color: Colors.grey)),
-      backgroundColor: Colors.grey.shade900,
-      content: Text(description, style: const TextStyle(color: Colors.grey)),
-      actions: [
-        TextButton(onPressed: () => Navigator.of(context).pop(), child: const Text('Cerrar', style: TextStyle(color: Colors.blueAccent)))
-      ],
-    );
   }
 }
